requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
django-model-utils==4.2.0
